+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - [OpenAPI] Add support for the `view:` option in Blueprinter parser.
 - Add `Rage::Daemon` (#339).
 - Enable non-blocking process monitoring (#341).
+- Allow to suppress FiberWrapper warnings (#347).
 
 ### Fixed